agrimonitor-backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
psycopg2-binary==2.9.6